refactor: now each monster form is added by the user
Restructure the code so that it allows more than 1 monster to be added.

Monster details like tempered/frenzied checkboxes are seperated in  another form.
Monster details form is added by clicking on a 'Add' button.

No more than 2 monsters allowed for now.

Selected monsters are part of an observable array wired to a handleChange method.