refactor!: make monster-form.js implement mediator
Now the module does not interact directly with the main create.js but with the mediator.

Instead of requiring variables from create.js, now the renamed monster-select-forms triggers events of the mediator

When the monster select forms change or get removed, the mediator event for monster change is sent and handled.